refactor: handleUpdateAttributes 함수 attributesUtils.js 파일로 이동 - 변경사유: handleUpdateAttributes 함수 renderElement.js, createElement__v2.js에서 모두 사용 가능하여 utils로 이동
