+++ src/utils/index.js
@@ -1,2 +1,2 @@
 export * from "./eventUtils.js";
-export * from "./camelToKebabUtil.js";
+export * from "./attributesUtils.js";
